refactor: Use SshMessage ctor arg for message ID instead of reflection
The current way of assigning Message IDs, i.e. using a `CLASS_ID_FIELD`
and accessing its value at runtime via reflection is not really ideal.
Reflection may fail at runtime instead of compile time, and makes it
harder to define custom `SshMessage` subclasses with values that are not
part of the `MessageIdConstant` enum.

This switches to a constructor-based approach, i.e. the message ID has
to be passed to the `SshMessage` constructor (either as a plain `byte` or
as a convenient `MessageIdConstant`).

Subclasses of `SshMessage` can still specify a parameter-less
constructor (that calls the `SshMessage` constructor with the proper
arguments internally), so this change does not prevent subclasses from
being proper JavaBeans.