fix(maven): Fix relative license path when used as a git submodule
Before this change, the `license-maven-plugin` requires the
`license_header_plain.txt` file in the same directory as the root
`pom.xml` rather than the `pom.xml` that enables the license plugin.
This fixes the issue by setting using the root directory of SSH-Attacker
project rather than Maven's multi-module project directory.

Unfortunately, mvn's dependency management solely relies on JAR releases
from Maven repositories instead of allowing pinning of specific commits
from Git repositories as dependencies like other build systems do (e.g.
Rust's cargo or Python's pip/setuptools).

This limitation does not play well with the way the *tls-attacker*
project develops its libraries, which are mainly used as in-development
snapshots rather than public releases. As a workaround, students are
required to *know* which commits from the different repositories they
need, build those projects and install them to their
local Maven repository in the right order. This is brittle, error-prone
and does not play well with CI builds.

As a workaround, it's possible to pin specific commits by adding
dependencies as a git submodule and adding those submodule as child
projects to the root `pom.xml`.

Without this fix, it always attempts to load `license_header_plain.txt`
from the root of the git repository, which may either fail because the
file is missing or result in using the wrong file (the file of the
including project instead of the one belonging to to the submodule).