fix(core): Fix writing of invalid XML if `Config` is subclassed
If `Config` is subclassed, using JAXB/Jakarta serialization will lead to a
broken XML representation:

    ConfigIO.write(someInstanceOfConfigSubclass, outputStream)

In the this case, XML root element will be the name of the subclass (or
whatever the `@XmlRootElement` annotation of the *subclass* says), not
`config` as intended. Also, additional fields of that subclass will be
included in the resulting XML unless these are annotated with
`@XmlTransient`.

Hence, the resulting XML files will be incompatible with plain
SSH-Attacker and unmarshalling such config objects will not work
anymore.

This commit fixes the issue by using the same JAXB context for unmarshalling
that we already use for marshalling.

See also: https://github.com/tls-attacker/SSH-Attacker/issues/183#issuecomment-1347067008