fix(core): Determine usable public key algorithms for authentication
Before, the `UserAuthPubkeyMessagePreparator` simply used the the **key
format** name as **algorithm** name. This works by change, because most
(all?) formats have a corresponding algorithm of the same name. Still,
this is incorrect.

This changes the behavior to the following, starting with the first user
key:

1. Check what key formats is used by the key.
2. Check what public keys algorithms support the same key format. If
   there are none, continue with the next user key.
3. Use the first key-and-algorithm-pair that is found.

Right now, this should not really make a difference, but it lays the
foundation for making the public key algorithms that are used for
authentication configurable.
This is useful because we may want to authenticate using an RSA key, but
not use the legacy `ssh-rsa` algorithm, but something more secure (e.g.
the algorithms from [RFC 8332, section 3][1]).

[1]: https://www.rfc-editor.org/rfc/rfc8332#section-3