fix(core): Return `false` in `SendAction::executedAsPlanned` on failure
If a failure occurred when sending issues (e.g. a write failure), then
`SendAction::executedAsPlanned` should return `false`. Currently, it
always returns `true` if it was executed.

This fix is just a band-aid. The actual fix would be to also make
`SendAction::getSendMessages` return the list of messages that were
actually sent (instead of just the list of messages that were *supposed
to be sent*), and then compare these two lists. Unfortunately, this
requires duplicating the whole JAXB serialization code that already
exists in `MessageAction` and `ReceiveAction`. Until a better solution
has been found, this simpler fix should suffice.