Implement parsing and serialization of data packets send over ChannelDataMessages.
Implemented DataPackets, that can be used for protocols that are running
on Channel Data (inluding AuthAgent). To provide same abstraction as
classic SSH packets, I implemented it using AbstractDataPacket and
provided PassThroughPacket, that really does nothing, and just pass
through the data. This adds unnecessery complexety, but provides
abstraction.

To provide same abstraction as the classic SSH packet layer I implemented an AbstractDataPacketLayer (incuding Layers:
DataPacketLayer and PassThroughPacketLayer). Tho this is not really
necessery it only adds more complexity and extra objects that needs to
be created. But maybe in future this abstraction is useful.

Implemented StringDataMessages so Shell / Exec Commands, can be more
easily created.

Implemented UnknownDataMessage for data on ChannelDataMessages, that
have no assosiated Parser.

[SFTP]
- Implemented basic Info and Version Message
- Added Client Version and Server Version (with default 3) to: Chooser, Config, DefaultChooser, SshContext

[ReceiveAction]
- Added IGNORE_CHANNEL_DATA_WRAPPER Option, to ignore the ChannelDataMessage, because its data is parsed to a new message

[WorkflowConfigurationFactory]
- Added addSftpInitActions()